+++ src/brevitas/export/onnx/qonnx/manager.py
@@ -1,24 +1,20 @@
 # Copyright (C) 2023, Advanced Micro Devices, Inc. All rights reserved.
 # SPDX-License-Identifier: BSD-3-Clause
 
-from functools import partial
 from typing import Optional
 from typing import Tuple
 from typing import Union
 
-from packaging.version import parse
 from torch import Tensor
 from torch.nn import Module
 
-from brevitas import torch_version
-from brevitas.export.inference.manager import _override_create_quant_tensor
 from brevitas.export.manager import _set_proxy_export_handler
 from brevitas.export.manager import _set_proxy_export_mode
 from brevitas.export.manager import _set_recurrent_layer_export_handler
 from brevitas.export.manager import _set_recurrent_layer_export_mode
 from brevitas.export.onnx.debug import DebugMarkerFunction
 from brevitas.export.onnx.manager import ONNXBaseManager
-from brevitas.graph.calibrate import QuantizationStatusManager
+from brevitas.export.onnx.manager import ONNXDynamoExportMixin
 from brevitas.quant_tensor import QuantTensor
 from brevitas.utils.logging import setup_logger
 
@@ -103,26 +99,7 @@ class QONNXManager(ONNXBaseManager):
             module, args, export_path, input_shape, input_t, disable_warnings, **onnx_export_kwargs)
 
 
-class QONNXDynamoManager(QONNXManager):
-    run_onnx_passes = False  # Skip the optimization step from onnxoptimizer. False required for QONNX + Dynamo to keep ONNX metadata
-    onnx_passes = ["eliminate_unused_initializer"]
-    custom_fns = []
-
-    @classmethod
-    def set_export_mode(cls, model: Module, enabled: bool):
-        super(QONNXDynamoManager, cls).set_export_mode(model=model, enabled=enabled)
-        # TODO: Move to a DynamoExport Mixin?
-        if enabled:
-            return_quant_tensor_state = QuantizationStatusManager.disable_return_quant_tensor(model)
-            disable_quant_tensor = partial(_override_create_quant_tensor, state=True)
-            model.apply(disable_quant_tensor)
-            model._brevitas_return_quant_tensor_state = return_quant_tensor_state  #  Store this state in the model, to re-enable later
-        else:
-            enable_quant_tensor = partial(_override_create_quant_tensor, state=False)
-            model.apply(enable_quant_tensor)
-            QuantizationStatusManager.restore_return_quant_tensor(
-                model, model._brevitas_return_quant_tensor_state)
-            del model._brevitas_return_quant_tensor_state
+class QONNXDynamoManager(ONNXDynamoExportMixin, QONNXManager):
 
     @classmethod
     def export_onnx(
@@ -134,15 +111,7 @@ class QONNXDynamoManager(QONNXManager):
             input_t: Optional[Union[Tensor, QuantTensor]],
             disable_warnings,
             **onnx_export_kwargs):
-        assert not parse("2.8") > torch_version, f"QONNX Export with `dynamo=True` only supported for PyTorch>=2.8. Current PyTorch version: {str(torch_version)}"
-        assert onnx_export_kwargs["dynamo"]
-        key = "optimize"
-        wrn_str = f"Optimize=True is recommended with QONNX export with dynamo=True"
-        if key in onnx_export_kwargs.keys():
-            if not onnx_export_kwargs[key]:
-                logging.warning(wrn_str)
-        else:
-            logging.warning(wrn_str)
+        cls._check_dynamo_export_kwargs(onnx_export_kwargs)
         model = super(QONNXDynamoManager, cls).export_onnx(
             module, args, export_path, input_shape, input_t, disable_warnings, **onnx_export_kwargs)
         return model
